blockchain: Move UTXO DB methods to separate file.
This moves the methods that deal with the UTXO database buckets from
chainio.go to a separate file, utxoio.go.

This is part of moving the UTXO set and state to a separate database.
